Split Timeout object into timeoutUtils file, addressed additional feedback
