+++ package.nls.json
@@ -12,7 +12,7 @@
     "AWS.configuration.description.logLevel": "The AWS Toolkit's log level (changes reflected on restart)",
     "AWS.configuration.description.onDefaultRegionMissing": "Action to take when a Profile's default region is hidden in the Explorer. Possible values:\n* `add` - shows region in the explorer\n* `ignore` - does nothing with the region\n* `prompt` - (default) asks the user what they would like to do.",
     "AWS.configuration.description.samcli.debug.attach.retry.maximum": "If the Toolkit is unable to attach a debugger, this is the number of times to retry before giving up.",
-    "AWS.configuration.description.samcli.debug.attach.timeout": "Maximum amount of time to connect a debugger to a SAM Local debug run (in milliseconds)",
+    "AWS.configuration.description.samcli.debug.attach.timeout": "Maximum amount of time to wait for debugger to connect to a SAM Local debug run (in milliseconds)",
     "AWS.configuration.description.samcli.location": "Location of SAM CLI. SAM CLI is used to create, build, package, and deploy Serverless Applications. [Learn More](https://aws.amazon.com/serverless/sam/)",
     "AWS.configuration.description.telemetry": "Enable anonymous usage data about the toolkit to be sent to AWS. `\"Use the same setting as VS Code\"` respects the value specified in \"`telemetry.enableTelemetry`\"",
     "AWS.telemetry.notificationMessage": "Please help improve the AWS Toolkit by enabling anonymous usage data to be sent to AWS. You can always change your mind later by going to the \"AWS Configuration\" section in your user settings",
@@ -135,7 +135,7 @@
     "AWS.samcli.local.invoke.ended": "Local invoke of SAM Application has ended.",
     "AWS.samcli.local.invoke.error": "Error encountered running local SAM Application",
     "AWS.samcli.local.invoke.port.not.open": "The debug port doesn't appear to be open. The debugger might not succeed when attaching to your SAM Application.",
-    "AWS.samcli.local.invoke.debugger.timeout": "The SAM process did not make the debugger available within the timelimit",
+    "AWS.samcli.local.invoke.debugger.timeout": "The SAM process did not make the debugger available within the time limit",
     "AWS.samcli.notification.version.invalid": "Your SAM CLI version {0} does not meet requirements ({1}\u00a0\u2264\u00a0version\u00a0<\u00a0{2}). {3}",
     "AWS.samcli.notification.version.valid": "Your SAM CLI version {0} is valid.",
     "AWS.samcli.recommend.update.toolkit": "Please update your AWS Toolkit.",
